fix(daos-list): refactored to single component with unified styles
