+++ astro_2.0/components/DaosList/index.ts
@@ -1,0 +1,1 @@
+export * from './DaosList';
